baykar_iha/requirements.txt
Django==4.2
djangorestframework==3.14
psycopg2-binary
djangorestframework-simplejwt==5.2
drf-yasg
django-cors-headers
